Added workflow to build pull-requests and trigger deployment (#77)
Build and deploy a pull-request by adding the 'preview' label to the pull request

Adding the label will build the image and update the image tag in the values.pr.yaml file on a branch in the gitops repo. The gitops repo looks after the deployment.

Removing the label or closing the pr will delete the branch. The gitops repo will then delete the deployment.

I did a bunch of refactoring of the workflow as part of this